Fixed bug when accessing attr in nested messages
The bug occured when trying to get/set an attribute in a message that is
nested inside another message (outer_msg.inner_msg.attribute). This was
being caused by _ftype_map not being set on the inner message. It seems
that super().__setattr__ on the outer message treats the inner message value
differently than primitives.